+++ package.json
@@ -72,7 +72,6 @@
     "@elizaos/plugin-openai": "workspace:*",
     "@hapi/shot": "^6.0.1",
     "@types/hapi": "^18.0.14",
-    "eliza": "git+https://github.com/elizaos/eliza.git#v2-develop",
     "node-opus": "^0.3.3",
     "opusscript": "^0.1.1",
     "vittest": "^1.0.2",
